[ADD] l10n_fr_account_vat_return_selenium
Move Selenium IDE support from l10n_fr_account_vat_return to a new
decidated module l10n_fr_account_vat_return_selenium.
This feature is beta, so I prefer to move it to a separate module.
