+++ l10n_fr_account_vat_return_selenium/README.rst
@@ -1,0 +1,1 @@
+Will be auto-generated from readme subdir
