+++ l10n_fr_account_vat_return_selenium/__init__.py
@@ -1,0 +1,1 @@
+from . import models
