+++ l10n_fr_account_vat_return_selenium/models/__init__.py
@@ -1,0 +1,1 @@
+from . import l10n_fr_account_vat_return
