+++ l10n_fr_account_vat_return_selenium/readme/CONTRIBUTORS.rst
@@ -1,0 +1,1 @@
+* Alexis de Lattre <alexis.delattre@akretion.com>
