+++ l10n_fr_account_vat_return_selenium/readme/DESCRIPTION.rst
@@ -1,0 +1,8 @@
+This module aims at providing an alternative in-between 2 solutions:
+
+* the teletransmission of VAT returns. Teletransmission of VAT return requires a paid subscription to a teletransmission gateway. Such a solution is provided by the OCA module **l10n_fr_account_vat_return_teledec**.
+* the manual copy of the CA3 form on impots.gouv.fr from the lines of the VAT return in Odoo (or the PDF report generated by Odoo). As Odoo developpers, we hate stupid manual work.
+
+This module is the only solution which proposes a sort of free tele-transmission of the CA3, via the auto-fill of the CA3 form on impots.gouv.fr. For that, this module relies on `Selenium IDE <https://www.selenium.dev/selenium-ide/>`_, which is a plugin available for Firefox and Chrome. Selenium is a free-software project initially designed to test Websites.
+
+This solution only works on the form 3310-CA3 ; it doesn't work when you have VAT boxes on the form 3310-A or you asked for a VAT credit reimbursement via form 3519.
