+++ l10n_fr_account_vat_return_selenium/readme/ROADMAP.rst
@@ -1,0 +1,1 @@
+* Move the code to a python lib, to mutualize the code across Odoo versions and give the opportunity to share the feature with other projects.
