+++ setup/l10n_fr_account_vat_return_selenium/odoo/addons/l10n_fr_account_vat_return_selenium
@@ -1,0 +1,1 @@
+../../../../l10n_fr_account_vat_return_selenium
